+++ mockserver/mockserver-core/src/test/java/org/mockserver/httpclient/netty/NettyHttpClientConnectionPoolTest.java
@@ -89,7 +89,8 @@ public class NettyHttpClientConnectionPoolTest {
     }
 
     private NettyHttpClient unpooledClient() {
-        return new NettyHttpClient(configuration(), mockServerLogger, clientEventLoopGroup, null, false);
+        Configuration configuration = configuration().forwardConnectionPoolEnabled(false);
+        return new NettyHttpClient(configuration, mockServerLogger, clientEventLoopGroup, null, false);
     }
 
     @Test